moved cli-staff to separate package
